Updated UI
Moved previous site files (now mostly unused, but preserved just to be safe) - now located in '_old' folder.

NEW UI - current site version, nearly complete. Work to be done on further reducing scale in the header and popups.

UI is finally functional and ready to be synced with the API calls.