Added OASIS API RELEASE HISTORY to root of OASIS API REST WEB API project.
This needs to be updated wheneber we do anywork that will effect the OASIS API (even changes to properties, etc on OASIS.API.CORE objects that are used in the API, etc...

Then when we do a release, this file will be linked to the Swagger documentation. We also then need to copy it to the WIKI and GitHub Documentation.